Nest xenophile's WebIdl model under the WebIdl namespace
`WebIdlDefinition`, `WebIdlParser`, and the three file-mates the definition
carried with it — `WebIdlArgument`, `WebIdlMember`, `WebIdlField` — plus
`WebIdlError`, all become members of `object WebIdl`. None was exported, so this
tidies xenophile's internal surface rather than the umbrella's.

`WebIdlDialect` is deliberately untouched: xenophile resolves a grammar by
`Class.forName(dialectType.typeSymbol.fullName)`, which only agrees with the JVM
binary name for a top-level object.

The parser needed `scala.collection.immutable.{::, Nil as SNil}`, which the
proposal warns re-points the cons pattern for everything it is in scope over.
It is scoped to `object Parser` rather than the file, so the rest of `WebIdl` is
unaffected. Verified with the full `xenophile.test.run` (203 passed), not just a
compile: nothing in `soundness.all` exercises the reflective lookup.
